# docs/swagger.yaml
definitions:
...
  exception.AppErr:
    properties:
      message:
        example: some error
        type: string
      status:
        example: 404
        type: integer
    type: object
  exception.ValidationError:
    properties:
      errors:
        items:
          $ref: '#/definitions/exception.validationErrorItem'
        type: array
      status:
        example: 400
        type: integer
    type: object
  exception.validationErrorItem:
    properties:
      key:
        example: email
        type: string
      message:
        example: email is invalid
        type: string
    type: object
...
  product.AdminProduct:
    properties:
      brand:
        $ref: '#/definitions/product.ProductBrand'
      category:
        $ref: '#/definitions/product.ProductCategory'
      description:
        example: Отлчиная куртка теплая
        type: string
      id:
        example: 1
        type: integer
      models:
        items:
          $ref: '#/definitions/product.AdminProductModelRelation'
        type: array
      slug:
        type: string
      title:
        example: Куртка теплая
        type: string
    type: object
  product.AdminProductModelRelation:
    properties:
      discount:
        type: integer
      id:
        example: 1
        type: integer
      image_path:
        type: string
      price:
        example: 15000
        type: integer
      product_id:
        type: integer
    type: object
  product.AdminProductResponse:
    properties:
      products:
        items:
          $ref: '#/definitions/product.AdminProduct'
        type: array
      total:
        type: integer
    type: object
  product.Color:
    properties:
      value:
        type: string
    type: object
  product.CreateProducModelImg:
    properties:
      img_path:
        type: string
      product_model_id:
        minimum: 1
        type: integer
    required:
    - img_path
    - product_model_id
    type: object
  product.CreateProductDto:
    properties:
      brand_id:
        example: 10
        minimum: 1
        type: integer
      category_id:
        example: 10
        minimum: 1
        type: integer
      description:
        example: Отлчиная куртка теплая
        minLength: 10
        type: string
      title:
        example: Куртка теплая
        minLength: 3
        type: string
    required:
    - brand_id
    - category_id
    - title
    type: object
  product.CreateProductModelDto:
    properties:
      discount:
        example: 10
        type: integer
      image_path:
        type: string
      price:
        example: 15000
        minimum: 1
        type: integer
      product_id:
        example: 10
        minimum: 1
        type: integer
    required:
    - price
    - product_id
    type: object
  product.Product:
    properties:
      brand:
        $ref: '#/definitions/product.ProductBrand'
      category:
        $ref: '#/definitions/product.ProductCategory'
      current_model:
        $ref: '#/definitions/product.ProductModel'
      description:
        example: Отлчиная куртка теплая
        type: string
      id:
        example: 1
        type: integer
      slug:
        type: string
      title:
        example: Куртка теплая
        type: string
    type: object
  product.ProductBrand:
    properties:
      id:
        example: 2
        type: integer
      img_path:
        type: string
      slug:
        example: verhnia-odezhda
        type: string
      title:
        example: adidas
        type: string
    type: object
  product.ProductCategory:
    properties:
      category_id:
        example: 3
        type: integer
      img_path:
        example: /static/example.webp
        type: string
      short_title:
        type: string
      slug:
        example: verhnia-odezhda
        type: string
      title:
        example: Верхняя одежда
        type: string
    type: object
  product.ProductModel:
    properties:
      discount:
        type: integer
      id:
        example: 1
        type: integer
      image_path:
        type: string
      images:
        items:
          $ref: '#/definitions/product.ProductModelImg'
        type: array
      options:
        items:
          $ref: '#/definitions/product.ProductModelOption'
        type: array
      price:
        example: 15000
        type: integer
      product_id:
        type: integer
      sizes:
        items:
          $ref: '#/definitions/product.ProductModelSize'
        type: array
    type: object
  product.ProductModelColors:
    properties:
      color:
        $ref: '#/definitions/product.Color'
      id:
        example: 1
        type: integer
      image_path:
        type: string
      slug:
        type: string
    type: object
  product.ProductModelImg:
    properties:
      id:
        example: 1
        type: integer
      img_path:
        type: string
    type: object
  product.ProductModelOption:
    properties:
      id:
        example: 4
        type: integer
      slug:
        example: color
        type: string
      title:
        example: Цвет
        type: string
      values:
        items:
          $ref: '#/definitions/product.ProductModelOptionValue'
        type: array
    type: object
  product.ProductModelOptionValue:
    properties:
      id:
        example: 44
        type: integer
      info:
        type: string
      value:
        example: Желтый
        type: string
    type: object
  product.ProductModelSize:
    properties:
      in_stock:
        type: integer
      literal:
        type: string
      model_id:
        type: integer
      size_id:
        type: integer
      size_model_id:
        type: integer
    type: object
  product.UpdateProductDto:
    properties:
      description:
        example: Отлчиная куртка теплая
        minLength: 10
        type: string
      title:
        example: Куртка теплая
        minLength: 3
        type: string
    required:
    - title
    type: object
  product.UpdateProductModelDto:
    properties:
      discount:
        example: 10
        minimum: 1
        type: integer
      image_path:
        type: string
      price:
        example: 15000
        minimum: 1
        type: integer
    required:
    - discount
    - price
    type: object
...
info:
  contact: {}
paths:
...
  /api/colored-models/:slug:
    get:
      consumes:
      - application/json
      description: Get color models
      parameters:
      - description: Product slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/product.ProductModelColors'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Get color models
      tags:
      - product
...
  /api/product/:
    post:
      consumes:
      - application/json
      description: Create product with dto
      parameters:
      - description: Create product dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/product.CreateProductDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Create product
      tags:
      - product
  /api/product/:id:
    delete:
      consumes:
      - application/json
      description: Delete product
      parameters:
      - description: product id parameter
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Delete product
      tags:
      - product
    patch:
      consumes:
      - application/json
      description: Update product
      parameters:
      - description: product id parameter
        in: path
        name: id
        required: true
        type: integer
      - description: update product dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/product.UpdateProductDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Update product
      tags:
      - product
  /api/product/add-photo:
    post:
      consumes:
      - application/json
      description: Add photo to product model
      parameters:
      - description: Add photo to product model
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/product.CreateProducModelImg'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Add photo to product model
      tags:
      - product
  /api/product/admin/products-list:
    get:
      consumes:
      - application/json
      description: Get products for admin panel
      parameters:
      - description: Page for pagination
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/product.AdminProductResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      security:
      - BearerToken: []
      summary: Get products for admin panel
      tags:
      - product
  /api/product/model:
    post:
      consumes:
      - application/json
      description: Create product model with dto
      parameters:
      - description: Create product model dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/product.CreateProductModelDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Create product-model
      tags:
      - product
  /api/product/model-full/:id:
    get:
      consumes:
      - application/json
      description: Get product model by model id
      parameters:
      - description: model id parameter
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/product.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Get product model by model id
      tags:
      - product
  /api/product/model/:id:
    delete:
      consumes:
      - application/json
      description: Delete product model
      parameters:
      - description: product model id parameter
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Delete product model
      tags:
      - product
    patch:
      consumes:
      - application/json
      description: Update product model
      parameters:
      - description: model id parameter
        in: path
        name: id
        required: true
        type: integer
      - description: update product-model dto
        in: body
        name: dto
        required: true
        schema:
          $ref: '#/definitions/product.UpdateProductModelDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Update product model
      tags:
      - product
  /api/product/remove-photo/:photoId:
    delete:
      consumes:
      - application/json
      description: Remove photo from product model
      parameters:
      - description: Photo id
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exception.ValidationError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/exception.AppErr'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/exception.AppErr'
      summary: Remove photo from product model
      tags:
      - product
...
swagger: "2.0"